+++ ledger/sandbox-perf/src/perf/benches/scala/com/digitalasset/platform/sandbox/perf/AcsBench.scala
@@ -15,7 +15,22 @@ import com.digitalasset.ledger.api.v1.event.CreatedEvent
 import com.digitalasset.ledger.api.v1.value.Identifier
 import com.digitalasset.ledger.client.services.acs.ActiveContractSetClient
 import com.digitalasset.platform.sandbox.services.TestCommands
-import org.openjdk.jmh.annotations.Benchmark
+import org.openjdk.jmh.annotations.{Benchmark, Level, Setup}
+
+class AcsBenchState extends PerfBenchState with DummyCommands with InfAwait {
+
+  def commandCount = 10000L
+
+  @Setup(Level.Invocation)
+  def submitCommands(): Unit = {
+    await(
+      dummyCreates(ledger.ledgerId)
+        .take(commandCount)
+        .mapAsync(100)(ledger.commandService.submitAndWait)
+        .runWith(Sink.ignore)(mat))
+    ()
+  }
+}
 
 class AcsBench extends TestCommands with InfAwait {
