+++ ledger/sandbox-perf/src/perf/benches/scala/com/digitalasset/platform/sandbox/perf/SimpleBench.scala
@@ -9,11 +9,14 @@ import akka.stream.scaladsl.Sink
 import com.digitalasset.daml.bazeltools.BazelRunfiles._
 import org.openjdk.jmh.annotations.Benchmark
 
+class SimpleBenchState extends PerfBenchState with DummyCommands with InfAwait
+
 class SimpleBench extends DummyCommands with InfAwait {
 
   override protected def darFile: File = new File(rlocation("ledger/test-common/Test-stable.dar"))
+
   @Benchmark
-  def ingest10kCommands(state: PerfBenchState): Unit = {
+  def ingest10kCommands(state: SimpleBenchState): Unit = {
     val commandCount = 10000L
     await(
       dummyCreates(state.ledger.ledgerId)
